requirements.txt
# Core dependencies
neo4j
python-dotenv
numpy
flask
sentence_transformers 
# Groq client for LLM
OPENAI
neo4j_graphrag
# Requests for API calls
requests
